# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/hub": {
            "post": {
                "description": "Create New Hub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hub"
                ],
                "summary": "Create New Hub",
                "operationId": "create-hub",
                "parameters": [
                    {
                        "description": "hub data",
                        "name": "jsonBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HubRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "include hub data in this response, see models.Hub",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "This api search hub and team information are matched with the key word (name, location, Type)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Search for hub/team information",
                "operationId": "search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search keyword",
                        "name": "key",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "include array of hub and team are matched the search key",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/team": {
            "post": {
                "description": "Create New Team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "Create New Team",
                "operationId": "create-team",
                "parameters": [
                    {
                        "description": "Team Type is one of (Platform, Fin, Music, Publishing, Studio), HubId is optional, maybe assign later",
                        "name": "jsonBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "include team data, see models.Team",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/team/assign-to-hub": {
            "post": {
                "description": "Team Hub Assign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "Team Hub Assign",
                "operationId": "assign-team-hub",
                "parameters": [
                    {
                        "description": "assigned data",
                        "name": "jsonBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HubAssign"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "new team data",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/user": {
            "post": {
                "description": "Create New User",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create New User",
                "operationId": "create-user",
                "parameters": [
                    {
                        "description": "Role is one of (Intern, S1, S2, S3, M1, M2, Head), team_id is optional, may be assign later",
                        "name": "jsonBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "include user data, see models.User",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/user/assign-to-team": {
            "post": {
                "description": "User Team Assign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "User Team Assign",
                "operationId": "assign-user-team",
                "parameters": [
                    {
                        "description": "assigned data",
                        "name": "jsonBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TeamAssign"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "new user data",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.HubAssign": {
            "type": "object",
            "properties": {
                "hub_id": {
                    "type": "integer"
                },
                "team_id": {
                    "type": "integer"
                }
            }
        },
        "models.HubRequest": {
            "type": "object",
            "properties": {
                "geo_location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.TeamAssign": {
            "type": "object",
            "properties": {
                "team_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.TeamRequest": {
            "type": "object",
            "properties": {
                "hub_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "team_type": {
                    "type": "string"
                }
            }
        },
        "models.UserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                }
            }
        }
    }
}